feat: ml left

+++ PerfumeTracker.Server/DTO/TagStatDTO.cs
@@ -1,3 +1,3 @@
 ﻿namespace PerfumeTracker.Server.Dto {
-	public record TagStatDto(int Id, string TagName, string Color, int Ml, int WornTimes);
+	public record TagStatDto(int Id, string TagName, string Color, decimal Ml, int WornTimes);
 }
